Bug quarantined in oil pipeline tests.
As suspected, the issue is a memory corruption of the "stack_" member
in the SimulationEngineRestart class. The error is a
"munmap_chunk(): invalid pointer:..." when calling the class dtor,
which can be caused by either:
- Some bad sequence of calls resulting in a double free() of stack_, or
- A reallocation at some point that invalidates the final free().

A bug report will be filed but no urgent action is needed,
because estimations can be carried out and the test aborts
after all calculi has finished, when the class is destructed.