Expected Success: fine-tune of internals.
New overriden tune() method to select the #(FE-sims) per iteration
and the #(steps) of each of these pilot runs.

We now also force the algorithm to select less thresholds than "optimal,"
using a fixed scale-down coefficient; see ThresholdsBuilderES.cpp:148.

TODO_0: use the test-suite to check the performance of the tune() choices
TODO_1: IDEM TODO_0 to check the scale-down with postprocess()