Design flaw for the refactor of ThresholdsBuilderES.
ThresholdsBuilderES uses now the fixed_effort() routine from
SimulationEnfineSFE to perform the pilot runs and choose thresholds.
However, a SimulationEngine requires an ImportanceFunction
*which already has thresholds* to run the simulations.

To break this circular dependency we will equip the SimulationEngine
with a field that marks it as "used for building thresholds."
Instances with this field in 'true' must tolerate the running of
simulations with an ImportanceFunction without thresholds.

TODO: refactor all SimulationEngine implementations to allow this.