Bugfix: ES works with compositional ifuns.
The issue was in the importance/thr.lvl duality.
We now hardcode the use of ImportanceValues for the thresholds-building
algorithms, disregarding the 'currentThresholds_' lvl information.

There is still a BUG when the expected success thresholds builder
is used in combination with the (simple) fixed effort engine.
The estimation is off, roughly 3x bigger than it should.
Suspiciously, 3 is the effort value of the last thresholds selected by ES.
Other algorithmic combinations, e.g. RESTART & ES or Fixed Effort & Hyb,
yield the proper estimate for transient tandem queue C=8.