Improvement on the previous two commits.
Now the Expected Success thresholds builder works on the general setting
(where rare events can happen at non-max importance values, e.g. database),
plus it uses again the information of the 'currentThresholds_' to steer
the internal Fixed Effort simulations.
This is much better than hardcoding the use of ImportanceValues,
because we can enforce "not-new paths to the rare event" to follow the
initially found path, with which we're currently building the thresholds.

The BUG remains re misestimation of the Fixed Effort engine
when used in combination with the Expected Success thresholds builder.
 ^^^ May be due to the thresholds vector interpretation:
     thresh[i] should be the effort to perform upon reaching level i.
Is this really how the Expected Success thresholds builder fills the vector?
The *efficiency* of RESTART when using ES may also be affected by all this.