Bugfix in SimulationEngineSFE::get_event_watcher().
The function defined a static variable, whose ctor arguments
implemented the member-function-binding functionality.
This means that on sucessive calls, the returned value
(i.e. the member function bound) was always the same.

Now the static variable declaration and its value assignemnt
are separate.

This could have been the cause behind SIGSEGVs in the test suite,
since the original object bound to the static variable could
be destroyed, and the function would then return a memory
address no longer owned by the process.