Syntax support for new property type "time bounded steady-state."
See the new class PropertyTBoundSS:
This is the same as the Rate properties, but where the batch size
(and also the transient phase) are explicitly defined by the user
as integer constants in the property.

This commits adds syntax support.
Next commit will add the simulation implementation for all engines.