Patched JANI to make IOSA- and Modest-compatible translations.
fig and The Modest Toolset support slightly different JANI formats.
Now "--to-jani" produces JANI that is compatible with fig. The output
matches mostly the one used in the JANI TACAS'17 paper (e9084441).

However, expressions reductions is now suppressed when translating
to JANI. This is also the default when translating from JANI, if the
model will not be evaluated but only translated ("--from-jani" switch).
Thus e.g. "const int A=1; const int B=A+1;" will not reduce "A+1" to 2.

Also, the new "--to-jani-modest" switch produces Modest-compatible JANI.
Expression reduction aside, the output produced matches the one used
for the STTT'19 article of the modes SMC tool (4deb98ad).