Testing the new probabilistic-branching capabilities.
Created a "forked tandem queue" to verify whether FIG is
implementing the branching correctly.
From Queue1 packets have probability 'p' of going to Queue2,
and with probability '1-p' they go to Queue3.
The rare event is still an overflow in Queue2; ignore Queue3.
Then the values of all properties in this model should be 'p' times
the value of the properties in the original tandem queue.