Bugfix (big and fat) for all estimations.
This seems to solve the misestimations for very precise CIs of the 2-tandem
queue Jackson, observed in the paper with Arnd for reproducing RESTART-Pj.
The fix consists of:

1. Deleting the --fast-math optimisation flag from RELEASE builds,
   which was enabling -funsafe-mast-optimisations in the background.
   [CMakeLists.txt:178]

2. Re-implementing all distributions for the clocks by hand,
   because re-creating C++ random deviates could be statistically unsound.
   [Clock.cpp]

BUG:
The excessive number of simulations required to converge in the aforementioned
case remains bad: more than 10x more samples than Jose and Arnd.