Resampling for Markovian models + New structure for Clock distributions.
Implemented resampling of clock values for exponential distributions.
This fixes the excessive #{samples} bug: we now equal Jose and Arnd.

Also changed the structure of the stochastic distributions in Clock.cpp
Now a distribution is a 'struct' with a 'sample()' method.
A new 'sample_cond()' method could be a generic way to implement resampling.