Clock value resampling (on split) for all distributions.
By default resampling is implemented via rejection sampling,
but bounded for an average of 4 rejections. If more rejections
are expected, resampling is discarded and Traial cloning keeps
the original clock values.

Special cases that implement resampling with the inverse method are
the exponential (heh), uniform, Weibull, and Rayleigh distributions.
This is more efficient than rejection sampling.

Added a battery of IOSAs in models/resampling_tests that demonstrate
unbiased estimation with the current resampling implementation.