Bugfixes in ThresholdsBuilder family.
For subclasses Fixed and SMC, do not set a threshold at the rare event
(i.e. the max importance value): since c5f9edb5 we do not split if the
event is rare, so having a threshold exactly at the rare event yields
under-estimations due to the lack of splitting.